feat(wallet): redesign homepage
