+++ agent-context/session-log.md
@@ -1,2 +1,6 @@
+
+## v0.2
+- Moved old wallet homepage to about section
+- Added new homepage content following style guide
 ## v0.1
 - Initial bootstrap: added agent-context folder and CI workflow.
